requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
backports.asyncio.runner==1.2.0
base58==2.1.1
bcrypt==5.0.0
bitarray==3.8.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
coincurve==20.0.0
crcmod==1.7
cryptography==46.0.3
cytoolz==1.1.0
Deprecated==1.3.1
dnspython==2.8.0
dramatiq==2.0.1
ecdsa==0.19.1
ed25519-blake2b==1.4.1
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
exceptiongroup==1.3.1
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
frozenlist==1.8.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
hdwallet==3.6.1
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.2
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pyunormalize==17.0.0
PyYAML==6.0.3
redis==7.1.0
regex==2026.1.15
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rlp==4.1.0
rsa==4.9.1
sentry-sdk==2.50.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.45
starlette==0.50.0
structlog==25.5.0
tomli==2.4.0
toolz==1.1.0
typer==0.21.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
web3==7.14.0
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
